api_hub/requirements.txt
azure-core==1.29.7
azure-cosmos==4.5.1
azure-functions==1.18.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
